Fix signed conversion warnings.

+++ test/hash_global_namespace_test.cpp
@@ -17,7 +17,7 @@ struct custom
 
     std::size_t hash() const
     {
-        return value_ * 10;
+        return static_cast<std::size_t>(value_ * 10);
     }
 
 #if !defined(BOOST_NO_ARGUMENT_DEPENDENT_LOOKUP)
